Change dir structure to split debugging and utility tools
