+++ eCP/include/eCP/debugging/debug_tools.hpp
@@ -1,12 +1,13 @@
-#ifndef UTILITY_H
-#define UTILITY_H
+#ifndef DEBUG_H
+#define DEBUG_H
 
 #include <vector>
 #include <eCP/index/eCP.hpp>
 
-namespace utility {
-
-std::vector<std::vector<float>> generate_descriptors(unsigned int count, unsigned int dimension, unsigned int upper_bound);
+/**
+ * Namespace contains functions to investigate the implementation and to help verify behaviour of the index.
+ */
+namespace debugging {
 
 void print_query_results(std::pair<std::vector<unsigned int>, std::vector<float>>& res, std::vector<float>& query, unsigned int k, const std::vector<Point>& S);
 
@@ -22,4 +23,4 @@ void print_index_levels(std::vector<Node>& root);
 
 }
 
-#endif // UTILITY_H
+#endif // DEBUG_H
